[refactor] 쿠폰 전체 및 상세 조회 메서드 경로 이동 ( #14 )
- CouponCommonController : ADMIN 과 USER 사용자들의 공용 컨트롤러이며, 쿠폰 전체 및 상세 조회 기능이 있다.
