+++ src/main/java/com/coupon/issuecouponservice/controller/coupon/CouponUserController.java
@@ -26,18 +26,6 @@ public class CouponUserController {
         couponService.issueCoupon(couponIssueParam, userDetails.getUser());
     }
 
-    // 쿠폰 전체 조회
-    @GetMapping("/coupon")
-    public List<CouponForm> readAllCoupons() {
-        return couponService.readAllCoupons();
-    }
-
-    // 쿠폰 상세 조회
-    @GetMapping("/coupon/{couponId}")
-    public CouponOneForm selectCoupon(@PathVariable Long couponId){
-        return couponService.selectCoupon(couponId);
-    }
-
     // 사용자 쿠폰 전체 조회
     @GetMapping("/{userId}/coupon")
     public List<CouponForm> readAllUserCoupons(@PathVariable Long userId){
